refactor: Rename "core" to "util"
